docs(javadoc): start documenting
* move internal classes
