+++ src/main/java/com/xenoterracide/gradle/semver/package-info.java
@@ -1,3 +1,6 @@
+/**
+ * This package contains the main classes for the plugin.
+ */
 @NullMarked
 package com.xenoterracide.gradle.semver;
